Admin menu components moved to admin folder
