+++ app/components/Menus/Admin/AdminCategoryPanelControl.latte
