+++ app/components/Menus/Admin/AdminPanelControl.latte
